Refactored class with exception handler class and moved interfaces to packages
