+++ social-app-backend/src/main/java/razepl/dev/socialappbackend/auth/AuthController.java
@@ -1,19 +1,18 @@
 package razepl.dev.socialappbackend.auth;
 
-import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import razepl.dev.socialappbackend.user.ServiceUser;
-import razepl.dev.socialappbackend.user.UserRepository;
-
-import java.util.stream.Collectors;
+import razepl.dev.socialappbackend.auth.interfaces.AuthInterface;
+import razepl.dev.socialappbackend.user.interfaces.ServiceUser;
+import razepl.dev.socialappbackend.user.interfaces.UserRepository;
 
 import static razepl.dev.socialappbackend.auth.constants.AuthMappings.AUTH_MAPPING;
 import static razepl.dev.socialappbackend.auth.constants.AuthMappings.REGISTER_MAPPING;
-import static razepl.dev.socialappbackend.auth.constants.AuthMessages.*;
+import static razepl.dev.socialappbackend.auth.constants.AuthMessages.ADDED_INFO;
+import static razepl.dev.socialappbackend.auth.constants.AuthMessages.SUCCESSFUL_REGISTER;
 import static razepl.dev.socialappbackend.constants.GlobalConstants.FRONTEND_ADDRESS;
 
 @Slf4j
@@ -32,17 +31,5 @@ public class AuthController implements AuthInterface {
         log.info(ADDED_INFO);
 
         return ResponseEntity.ok(SUCCESSFUL_REGISTER);
-    }
-
-    @Override
-    @ExceptionHandler(ConstraintViolationException.class)
-    public final ResponseEntity<String> handleValidationExceptions(ConstraintViolationException exception) {
-        String errorMessage = exception.getConstraintViolations().stream()
-                .map(error -> String.format(ERROR_FORMAT, error.getPropertyPath(), error.getMessage()))
-                .collect(Collectors.joining(ERROR_DELIMITER));
-
-        log.error(errorMessage);
-
-        return ResponseEntity.badRequest().body(errorMessage);
     }
 }
