+++ social-app-backend/src/main/java/razepl/dev/socialappbackend/user/interfaces/UserRepository.java
@@ -1,4 +1,4 @@
-package razepl.dev.socialappbackend.user;
+package razepl.dev.socialappbackend.user.interfaces;
 
 import org.springframework.data.jpa.repository.JpaRepository;
